Ensemble_Pred_BP/requirements.txt
absl-py==2.1.0
albumentations==1.0.0
asttokens==2.4.1
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.12.3
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools==5.3.3
certifi @ file:///croot/certifi_1707229174982/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
comm==0.2.2
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
filelock==3.13.4
flatbuffers==24.3.25
gast==0.4.0
gdown==5.1.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==2.10.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.34.0
importlib_metadata==7.1.0
ipykernel==6.26.0
ipython==8.12.3
ipywidgets==8.1.1
jedi==0.19.1
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
mat73==0.59
matplotlib==3.1.2
matplotlib-inline==0.1.6
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mtcnn==0.1.1
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.22.0
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.0
pandas==1.1.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1707233021655/work
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.2
requests @ file:///croot/requests_1707355572290/work
requests-oauthlib==2.0.0
retina-face==0.0.13
retinaface-pytorch==0.0.8
rsa==4.9
scikit-image==0.17.2
scikit-learn==1.0.2
scipy==1.5.2
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboardX==2.4.1
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2023.7.10
torch==1.9.0
torchaudio==0.12.1
torchvision==0.10.0
tornado==6.4
tqdm==4.64.0
traitlets==5.14.2
typing_extensions==4.5.0
urllib3 @ file:///croot/urllib3_1707770551213/work
wcwidth==0.2.13
Werkzeug==3.0.2
widgetsnbextension==4.0.10
wrapt==1.16.0
yacs==0.1.8
zipp==3.18.1
